Few things
Added sprite to spawner
Fixed bug with entering byte data for entities
Added option to get the tile id on right click

+++ src/designer/EntityExtrasFrame.java
@@ -105,7 +105,7 @@ public class EntityExtrasFrame extends JFrame implements ActionListener {
 		}
 
 		public byte getExtra2() {
-			return Byte.parseByte(extra2.getText());
+			return (byte) Integer.parseInt(extra2.getText());
 		}
 
 	}
